Reorganize rendering files (separate Vulkan from generic rendering)
Reduce 64bit to 32bit integer conversion warnings.
