+++ BRenderer/CMakeLists.txt
@@ -69,14 +69,14 @@ set(BRenderer_SOURCE_FILES
     "Importer/Importer.cpp"
     
     "Renderer/Allocators/StagingAllocator.cpp"
+    "Renderer/Vulkan/VkInitializerHelper.cpp" 
+    "Renderer/Vulkan/VmaImpl.cpp"
+    "Renderer/Vulkan/VulkanRenderDevice.cpp"
     "Renderer/Descriptors.cpp"
     "Renderer/DeviceBuffer.cpp"
     "Renderer/DevicePipeline.cpp"
-    "Renderer/VulkanRenderDevice.cpp"
     "Renderer/Shader.cpp" 
     "Renderer/Swapchain.cpp"
-    "Renderer/VkInitializerHelper.cpp" 
-    "Renderer/Vulkan/VmaImpl.cpp"
     
     "Files/FilesUtils.cpp" 
   )
@@ -84,13 +84,14 @@ set(BRenderer_SOURCE_FILES
 set(BRenderer_HEADER_FILES
     "BRenderer.h" 
     
-    "Core/App.h"
-    "Core/LogSystem.h"
-    "Core/PerspectiveCamera.h"
     "Core/Threading/Barrier.h" 
     "Core/Threading/Threading.h" 
     "Core/Threading/ThreadPool.h" 
     "Core/Threading/Work.h"
+    "Core/App.h"
+    "Core/LogSystem.h"
+    "Core/PerspectiveCamera.h"
+    "Core/thirdpartiesInc.h"
     
     "Geometry/Geometry.h"
     
@@ -112,17 +113,19 @@ set(BRenderer_HEADER_FILES
     "Importer/Importer.h"
     
     "Renderer/Allocators/StagingAllocator.h"
+    "Renderer/Vulkan/VkInitializerHelper.h"
+    "Renderer/Vulkan/VulkanInc.h"
+    "Renderer/Vulkan/VulkanRenderDevice.h"
     "Renderer/Descriptors.h"
     "Renderer/DeviceBuffer.h"
     "Renderer/DevicePipeline.h"
     "Renderer/RenderDefs.h"
-    "Renderer/VulkanRenderDevice.h"
     "Renderer/Shader.h"
     "Renderer/Swapchain.h"
-    "Renderer/VkInitializerHelper.h" 
+   
     
     "Files/FilesUtils.h"
-   "Core/thirdpartiesInc.h" "Renderer/VulkanInc.h") 
+) 
 
 source_group(Renderer REGULAR_EXPRESSION "Renderer/.*")
 source_group(Visualization REGULAR_EXPRESSION "Visualization/.*")
