+++ BRenderer/Renderer/Swapchain.h
@@ -1,7 +1,7 @@
 #ifndef BRR_SWAPCHAIN_H
 #define BRR_SWAPCHAIN_H
 
-#include <Renderer/VulkanInc.h>
+#include <Renderer/Vulkan/VulkanInc.h>
 #include <Renderer/RenderDefs.h>
 
 namespace brr::vis
